Conflicts:
	plugins/image/io/Exr/src/writer/EXRWriterProcess.hpp
	plugins/image/io/FFMpeg/src/reader/FFMpegReaderDefinitions.hpp
	plugins/image/io/FFMpeg/src/writer/FFMpegWriterDefinitions.hpp
	plugins/image/io/FFMpeg/src/writer/FFMpegWriterProcess.hpp
	plugins/image/process/color/BasicKeyer/Resources/tuttle.basickeyer.png
	plugins/image/process/color/CTL/Resources/tuttle.ctl.svg
	plugins/image/process/color/ColorDistribution/Resources/fr.tuttle.colordistribution.png
	plugins/image/process/color/ColorDistribution/Resources/fr.tuttle.colordistribution.svg
	plugins/image/process/color/ColorGradation/Jamfile.v2
	plugins/image/process/color/ColorGradation/Resources/fr.tuttle.colorgradation.png
	plugins/image/process/color/ColorGradation/Resources/fr.tuttle.colorgradation.svg
	plugins/image/process/color/ColorGradation/src/mainEntry.cpp
	plugins/image/process/color/Colorspace/SConscript
	plugins/image/process/color/Colorspace/src/ColorSpaceProcess.tcc
	plugins/image/process/color/HistogramKeyer/src/HistogramKeyerProcess.tcc
	plugins/image/process/filter/Blur/src/BlurProcess.tcc
	plugins/image/process/filter/FloodFill/src/FloodFillAlgorithm.hpp
	plugins/image/process/filter/Sobel/src/SobelProcess.tcc

+++ libraries/tuttle/src/tuttle/plugin/color/GradationProcess.tcc
@@ -30,7 +30,7 @@ struct computeFromSRGB
 	{
 		if( src > 0.04045 )
 		{
-			dst = DstChannel( pow( ( 1.0*src + 0.055 ) / 1.055 , 2.4 ) );
+			dst = DstChannel( pow( (double) ( 1.0*src + 0.055 ) / 1.055 , 2.4 ) );
 		}
 		else
 		{
@@ -46,7 +46,7 @@ struct computeToSRGB
 	{
 		if( src > 0.0031308 )
 		{
-			dst = DstChannel( 1.055 * pow( src , 1.0 / 2.4 ) - 0.055 );
+			dst = DstChannel( 1.055 * pow( (double) src , 1.0 / 2.4 ) - 0.055 );
 		}
 		else
 		{
@@ -66,10 +66,10 @@ struct computeFromCineon
 	template <typename SrcChannel, typename DstChannel>
 	void operator()( SrcChannel& src, DstChannel& dst ) const
 	{
-		//double tmpValue = pow( 10 , ( _blackPoint - _whitepoint ) / 300 );
-		//dst = ( pow( 10 , ( src*1024 - _whitePoint ) / 300 ) - tmpValue ) / tmpValue;
+		//double tmpValue = pow( (double) 10 , ( _blackPoint - _whitepoint ) / 300 );
+		//dst = ( pow( (double) 10 , ( src*1024 - _whitePoint ) / 300 ) - tmpValue ) / tmpValue;
 
-		dst = 1.010915615730753 * ( pow( 10 ,( 1023 * src - 685 ) / 300 ) - 0.010797751623277 );
+		dst = 1.010915615730753 * ( pow( (double) 10 ,(double) ( 1023 * src - 685 ) / 300 ) - 0.010797751623277 );
 	}
 };
 
@@ -84,10 +84,10 @@ struct computeToCineon
 	template <typename SrcChannel, typename DstChannel>
 	void operator()( SrcChannel& src, DstChannel& dst ) const
 	{
-		//double tmpValue = pow( 10 , ( _blackPoint - _whitepoint ) / 300 );
-		//dst = ( pow( 10 , ( src*1024 - _whitePoint ) / 300 ) - tmpValue ) / tmpValue;
+		//double tmpValue = pow( (double) 10 , ( _blackPoint - _whitepoint ) / 300 );
+		//dst = ( pow( (double) 10 , ( src*1024 - _whitePoint ) / 300 ) - tmpValue ) / tmpValue;
 
-		dst = 0.00042453028534042214 * ( 300 * log( 0.98920224837672 * src + 0.010797751623277 ) + 1577.270788700921 );
+		dst = 0.00042453028534042214 * ( 300 * log( (double) 0.98920224837672 * src + 0.010797751623277 ) + 1577.270788700921 );
 	}
 };
 
@@ -105,8 +105,8 @@ struct computeFromGamma
 		if( src > 0.0 )
 		{
 			// compute gamma value
-			//dst = DstChannel( pow( 10 , _value * log10( src ) ) );
-			dst = DstChannel( pow( src , _value ) );
+			//dst = DstChannel( pow( (double) 10 , _value * log10( (double) src ) ) );
+			dst = DstChannel( pow( (double) src , _value ) );
 		}
 		else
 		{
@@ -130,7 +130,7 @@ struct computeToGamma
 		if( src > 0.0 )
 		{
 			// compute gamma value
-			dst = DstChannel( pow( src , 1.0 / _value ) );
+			dst = DstChannel( pow( (double) src , 1.0 / _value ) );
 		}
 		else
 		{
@@ -146,7 +146,7 @@ struct computeFromPanalog
 	template <typename SrcChannel, typename DstChannel>
 	void operator()( SrcChannel& src, DstChannel& dst ) const
 	{
-		dst = ( pow ( 10, ( 1023.0 * src - 681.0 ) / 444.0 ) - 0.0408 ) / ( 1.0 - 0.0408 );
+		dst = ( pow ( (double) 10, ( 1023.0 * src - 681.0 ) / 444.0 ) - 0.0408 ) / ( 1.0 - 0.0408 );
 	}
 };
 
@@ -155,7 +155,7 @@ struct computeToPanalog
 	template <typename SrcChannel, typename DstChannel>
 	void operator()( SrcChannel& src, DstChannel& dst ) const
 	{
-		dst = ( 444.0 * log10( 0.0408 + ( 1 - 0.0408 ) * src ) + 681.0 ) / 1023.0;
+		dst = ( 444.0 * log10( (double) 0.0408 + ( 1 - 0.0408 ) * src ) + 681.0 ) / 1023.0;
 	}
 };
 
@@ -164,7 +164,7 @@ struct computeFromRedLog
 	template <typename SrcChannel, typename DstChannel>
 	void operator()( SrcChannel& src, DstChannel& dst ) const
 	{
-		dst = ( pow ( 10, ( 1023.0 * src - 1023.0 ) / 511.0 ) - 0.01 ) / ( 1.0 - 0.01 );
+		dst = ( pow ( (double) 10, ( 1023.0 * src - 1023.0 ) / 511.0 ) - 0.01 ) / ( 1.0 - 0.01 );
 	}
 };
 
@@ -173,7 +173,7 @@ struct computeToRedLog
 	template <typename SrcChannel, typename DstChannel>
 	void operator()( SrcChannel& src, DstChannel& dst ) const
 	{
-		dst = ( 511.0 * log10( 0.01 + ( 1 - 0.01 ) * src ) + 1023.0 ) / 1023.0;
+		dst = ( 511.0 * log10( (double) 0.01 + ( 1 - 0.01 ) * src ) + 1023.0 ) / 1023.0;
 	}
 };
 
@@ -182,7 +182,7 @@ struct computeFromViperLog
 	template <typename SrcChannel, typename DstChannel>
 	void operator()( SrcChannel& src, DstChannel& dst ) const
 	{
-		dst = pow ( 10, ( 1023.0 * src - 1023.0 ) / 500.0 );
+		dst = pow ( (double) 10, ( 1023.0 * src - 1023.0 ) / 500.0 );
 	}
 };
 
@@ -191,7 +191,7 @@ struct computeToViperLog
 	template <typename SrcChannel, typename DstChannel>
 	void operator()( SrcChannel& src, DstChannel& dst ) const
 	{
-		dst = ( 500.0 * log10( src ) + 1023.0 ) / 1023.0;
+		dst = ( 500.0 * log10( (double) src ) + 1023.0 ) / 1023.0;
 	}
 };
